chore: decorator in shared, remove re-exports (#191)
- shared: add PatternDecoratorConfig/Manager, runPatternConfigs/runPatternFromModel, DecoratorGenerator/Manager
- editor-view-dom: remove re-export files (decorator-manager, remote-decorator-manager, pattern-decorator-config-manager, decorator-generator); import DecoratorManager, RemoteDecoratorManager, PatternDecoratorConfigManager, DecoratorGeneratorManager from @barocss/shared; decorator/index exports only DOM-specific (types, registry, prebuilder, renderer, dom-query, position-calculator)
- editor-view-react: remove decorator/index re-export; import from @barocss/shared; drop decorator re-exports from package index
- tests: use @barocss/shared for DecoratorGenerator/Context in editor-view-dom integration tests
