--- packages/editor-view-dom/src/decorator/decorator-manager.ts
@@ -1,5 +1,0 @@
-/**
- * Re-export DecoratorManager from shared. editor-view-dom passes DecoratorRegistry as validator.
- */
-
-export { DecoratorManager } from '@barocss/shared';
