+++ packages/editor-view-dom/src/decorator/index.ts
@@ -1,14 +1,11 @@
 /**
- * Decorator 시스템 Export
+ * Decorator 시스템 Export (DOM 전용 + types).
+ * DecoratorManager, RemoteDecoratorManager, PatternDecoratorConfigManager, DecoratorGeneratorManager 는 @barocss/shared 에서 import.
  */
 
 export * from './types';
 export * from './decorator-registry';
-export * from './decorator-manager';
-export * from './remote-decorator-manager';
 export { DecoratorRenderer } from './decorator-renderer';
 export { DecoratorPrebuilder, type DecoratorModel } from './decorator-prebuilder';
 export { DOMQuery } from './dom-query';
 export { PositionCalculator, type DecoratorPosition } from './position-calculator';
-export { PatternDecoratorConfigManager, type PatternDecoratorConfig } from './pattern-decorator-config-manager';
-export { DecoratorGeneratorManager, type DecoratorGenerator, type DecoratorGeneratorContext } from './decorator-generator';
