--- packages/editor-view-dom/src/decorator/remote-decorator-manager.ts
@@ -1,5 +1,0 @@
-/**
- * Re-export RemoteDecoratorManager from shared.
- */
-
-export { RemoteDecoratorManager, type DecoratorOwner } from '@barocss/shared';
