+++ packages/editor-view-dom/test/integration/pattern-custom-decorator-render.test.ts
@@ -13,7 +13,7 @@ import { Editor } from '@barocss/editor-core';
 import { EditorViewDOM } from '../../src/editor-view-dom';
 import { DataStore } from '@barocss/datastore';
 import { define, element, slot, data, getGlobalRegistry, defineDecorator } from '@barocss/dsl';
-import type { DecoratorGenerator, DecoratorGeneratorContext } from '../../src/decorator/decorator-generator';
+import type { DecoratorGenerator, DecoratorGeneratorContext } from '@barocss/shared';
 import type { ModelData } from '@barocss/dsl';
 import { expectHTML } from '../utils/html';
