--- packages/editor-view-react/src/decorator/index.ts
@@ -1,10 +1,0 @@
-export {
-  PatternDecoratorConfigManager,
-  type PatternDecoratorConfig,
-} from './pattern-decorator-config-manager.js';
-export { runPatternConfigs, runPatternFromModel } from './run-pattern.js';
-export {
-  DecoratorGeneratorManager,
-  type DecoratorGenerator,
-  type DecoratorGeneratorContext,
-} from './decorator-generator.js';
